app/requirements.txt
numpy==2.0.2
pandas==2.2.3
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.2
urllib3==2.2.3
SQLAlchemy==2.0.36
Flask
gunicorn